Switch all C source files to C++.
There's no reason to avoid C++.  We were in fact already using it in a
few places.  And with C++ we can avoid the awkward and error-prone C
string manipulation functions, among other things.

+++ src/catchsegv/catchsegv.cpp
@@ -83,7 +83,7 @@ EnumWindowCallback(HWND hWnd, LPARAM lParam)
     DWORD dwThreadId;
 
     dwThreadId = GetWindowThreadProcessId(hWnd, &dwProcessId);
-    if (dwProcessId == lParam) {
+    if (dwProcessId == (DWORD)lParam) {
         if (GetWindowLong(hWnd, GWL_STYLE) & DS_MODALFRAME) {
             char szWindowText[256];
             if (GetWindowTextA(hWnd, szWindowText, _countof(szWindowText)) <= 0) {
